# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ProjectOS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List all team members
      responses:
        "200":
          description: Members list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
    post:
      operationId: createMember
      tags: [members]
      summary: Create a team member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMember"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /members/{id}:
    patch:
      operationId: updateMember
      tags: [members]
      summary: Update a team member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMember"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
...
  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: List goals with key results
      responses:
        "200":
          description: Goals list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGoal"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update a goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGoal"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List announcements
      parameters:
        - name: projectId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Announcements list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create an announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnnouncement"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{id}:
    patch:
      operationId: updateAnnouncement
      tags: [announcements]
      summary: Update an announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAnnouncement"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{id}/react:
    post:
      operationId: reactToAnnouncement
      tags: [announcements]
      summary: Add a reaction to an announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                emoji:
                  type: string
              required:
                - emoji
      responses:
        "200":
          description: Reaction added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{id}/comments:
    post:
      operationId: commentOnAnnouncement
      tags: [announcements]
      summary: Add a comment to an announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                authorId:
                  type: integer
              required:
                - text
                - authorId
      responses:
        "200":
          description: Comment added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents
      parameters:
        - name: projectId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Documents list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create a document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocument"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    patch:
      operationId: updateDocument
      tags: [documents]
      summary: Update a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDocument"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Member:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        initials:
          type: string
        color:
          type: string
        role:
          type: string
        rate:
          type: number
        capacity:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - initials
        - color
        - role
        - rate
        - capacity
        - createdAt
    CreateMember:
      type: object
      properties:
        name:
          type: string
        initials:
          type: string
        color:
          type: string
        role:
          type: string
        rate:
          type: number
        capacity:
          type: integer
      required:
        - name
        - initials
        - color
        - role
        - rate
        - capacity
    UpdateMember:
      type: object
      properties:
        name:
          type: string
        initials:
          type: string
        color:
          type: string
        role:
          type: string
        rate:
          type: number
        capacity:
          type: integer
...
    KeyResult:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        progress:
          type: integer
        target:
          type: number
        current:
          type: number
        unit:
          type: string
      required:
        - id
        - title
        - progress
        - target
        - current
        - unit
    Goal:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        status:
          type: string
          enum: [on_track, at_risk, off_track]
        progress:
          type: integer
        due:
          type: string
          format: date-time
        ownerId:
          type: integer
        projectId:
          type: ["integer", "null"]
        keyResults:
          type: array
          items:
            $ref: "#/components/schemas/KeyResult"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - status
        - progress
        - due
        - ownerId
        - keyResults
        - createdAt
    CreateGoal:
      type: object
      properties:
        title:
          type: string
        status:
          type: string
          enum: [on_track, at_risk, off_track]
        progress:
          type: integer
        due:
          type: string
          format: date-time
        ownerId:
          type: integer
        projectId:
          type: ["integer", "null"]
        keyResults:
          type: array
          items:
            type: object
            properties:
              title:
                type: string
              progress:
                type: integer
              target:
                type: number
              current:
                type: number
              unit:
                type: string
            required:
              - title
              - target
              - current
              - unit
      required:
        - title
        - due
        - ownerId
    UpdateGoal:
      type: object
      properties:
        title:
          type: string
        status:
          type: string
          enum: [on_track, at_risk, off_track]
        progress:
          type: integer
        due:
          type: string
          format: date-time
        keyResults:
          type: array
          items:
            type: object
            properties:
              id:
                type: ["integer", "null"]
              title:
                type: string
              progress:
                type: integer
              target:
                type: number
              current:
                type: number
              unit:
                type: string
            required:
              - title
              - target
              - current
              - unit
    Comment:
      type: object
      properties:
        authorId:
          type: integer
        text:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
        - authorId
        - text
        - timestamp
    Announcement:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        authorId:
          type: integer
        projectId:
          type: ["integer", "null"]
        pinned:
          type: boolean
        reactions:
          type: object
          additionalProperties:
            type: integer
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - content
        - authorId
        - pinned
        - reactions
        - comments
        - createdAt
    CreateAnnouncement:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        authorId:
          type: integer
        projectId:
          type: ["integer", "null"]
        pinned:
          type: boolean
      required:
        - title
        - content
        - authorId
    UpdateAnnouncement:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        pinned:
          type: boolean
    DocumentVersion:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        authorId:
          type: integer
        label:
          type: string
      required:
        - timestamp
        - authorId
        - label
    Document:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        icon:
          type: string
        projectId:
          type: ["integer", "null"]
        authorId:
          type: integer
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        pinned:
          type: boolean
        versions:
          type: array
          items:
            $ref: "#/components/schemas/DocumentVersion"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - icon
        - authorId
        - content
        - tags
        - pinned
        - versions
        - createdAt
        - updatedAt
    CreateDocument:
      type: object
      properties:
        title:
          type: string
        icon:
          type: string
        projectId:
          type: ["integer", "null"]
        authorId:
          type: integer
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        pinned:
          type: boolean
      required:
        - title
        - icon
        - authorId
        - content
    UpdateDocument:
      type: object
      properties:
        title:
          type: string
        icon:
          type: string
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        pinned:
          type: boolean
...