# lib/api-zod/src/generated/types/createAnnouncement.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ProjectOS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAnnouncement {
  title: string;
  content: string;
  authorId: number;
  /** @nullable */
  projectId?: number | null;
  pinned?: boolean;
}